fixed project errors and inconsistencies from branch merges
