+++ directKickShop.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 16.0.29806.167
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "dropShippingApp", "dropShippingApp\dropShippingApp.csproj", "{2B8E851D-2FCE-4186-A93C-DC70CE7E67EC}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UnitTests", "UnitTests\UnitTests.csproj", "{7E7388E4-F869-4268-81D5-0FB5B2128206}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +17,10 @@ Global
 		{2B8E851D-2FCE-4186-A93C-DC70CE7E67EC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2B8E851D-2FCE-4186-A93C-DC70CE7E67EC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2B8E851D-2FCE-4186-A93C-DC70CE7E67EC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7E7388E4-F869-4268-81D5-0FB5B2128206}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7E7388E4-F869-4268-81D5-0FB5B2128206}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7E7388E4-F869-4268-81D5-0FB5B2128206}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7E7388E4-F869-4268-81D5-0FB5B2128206}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
